Change ``/replays/most-viewed`` to ``/most-viewed`` for Scores
